Move bokeh min files to src/vendor and change imports
